time_clock/requirements.txt
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
fcm-django==1.0.5
firebase-admin==5.2.0
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
msgpack==1.0.3
mysqlclient==2.0.1
Naked==0.1.31
openapi-codec==1.3.2
packaging==21.3
Pillow==8.3.2
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
serializers==0.2.4
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
status==0.2.5
uritemplate==3.0.1
urllib3==1.26.7
